% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCAcompare.R
\name{CCAcompare}
\alias{CCAcompare}
\title{CCAcompare}
\usage{
CCAcompare(ref_mat, expr_mat, do_plot = FALSE, cor_function_use = cor,
  ...)
}
\arguments{
\item{ref_mat}{A variable x observation (i.e. genes as rows, samples as
columns) matrix of reference values}

\item{expr_mat}{A variable x observation matrix of experimental values}

\item{do_plot}{Plot the correlation matrix. Default: FALSE}

\item{cor_function_use}{Correlation function to be used.
Default: stats::cor}

\item{...}{Additional parameters to pass to the correlation function}
}
\value{
A matrix of correlation values
}
\description{
Compares the an experimental matrix of expression values against a reference
matrix of expression values to that of an entry in the Gene Expression
Omnibus database.
}
\details{
Performs a canonical correlation analysis on the two datasets and then
assesses the correlation using the CC scores.
}
