% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CancerProtein.R
\name{compareCancerProtein}
\alias{compareCancerProtein}
\title{Comparison of protein expression between normal and cancer tissues}
\usage{
compareCancerProtein(cancer_type)
}
\arguments{
\item{cancer_type}{A character string specifying the cancer type of interest.
The value must be one of the valid cancer names available in
\code{cancer_tissue_map$cancer}. If an invalid cancer type is provided,
the function will stop execution and throw an error.}
}
\value{
A list containing two components:
\itemize{
\item \code{table} - A tibble of metadata with the following columns:
\describe{
\item{ensembl}{Ensembl gene identifier.}
\item{gene}{Gene symbol.}
\item{normal_rank}{Average protein expression rank in normal tissue.}
\item{cancer_rank}{Weighted protein expression rank in cancer tissue.}
\item{delta_rank}{Difference (\eqn{cancer\_rank - normal\_rank}).}
\item{direction}{Direction of change ("Up", "Down", or "No change").}
}
\item \code{plot} - A \pkg{ggplot2} bar plot showing the direction and
magnitude of protein rank changes for each gene.
The down regulated genes are labeled in red, the up regulated genes are labeled in blue,
and the genes with no change in protein expression are labeled in green.

}
}
\description{
This function compares the protein expression levels between normal and
pathology (cancer) tissues for a specified cancer type. It calculates the
change in expression rank and direction ("Up", "Down", or
"No change") for each gene, and visualizes the shift in expression.
}
\details{
The function uses the \pkg{HPAanalyze} dataset (\code{hpaDownload}) for
expression data, and the internal dataset \code{\link{cancer_tissue_map}}
to match each cancer type to its corresponding tissue.
}
\examples{
\dontrun{
# Example 1: Compare protein expression in a common tissue

output <- compareCancerProtein(cancer_type = "breast cancer")
output$table
output$plot
}

}
\references{
Tran AN, Dussaq AM, Kennell Jr T, Willey CD, Hjelmeland AB (2019).
"HPAanalyze: an R package that facilitates the retrieval and analysis of the
Human Protein Atlas data." MC Bioinformatics 20, 463 (2019).
\url{https://doi.org/10.1186/s12859-019-3059-z}

Wickham H (2016). ggplot2: Elegant Graphics for Data Analysis.
Springer-Verlag New York. ISBN 978-3 319-24277-4, \url{https://ggplot2.tidyverse.org.}

Wickham H, Francois R, Henry L, Muller K, Vaughan D (2025).
dplyr: A Grammar of Data Manipulation. R package version 1.1.4,
\url{https://dplyr.tidyverse.org.}
}
