% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Correlation.R
\name{compute_correlation}
\alias{compute_correlation}
\title{Converts protein expression data queried from HPAanalyze, to numeric values}
\usage{
compute_correlation(gene_NAMES = NULL, tissue_NAMES = NULL)
}
\description{
A function that converts the protein expression data of a given gene/tissue
from HPAanalyze, into numeric values. HPAnalyze returns protein expression as
strings, for e.g.("low", "medium", etc). In order to determine the pearson
correlation between the RNA expression and protein expression, we convert these
strings into a numeric value.
}
