% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{convert_to_gtex}
\alias{convert_to_gtex}
\title{Convert HPA tissue names to GTEx tissue names}
\usage{
convert_to_gtex(tissues)
}
\arguments{
\item{tissues}{Character vector of HPA tissue names.
Valid tissue names are listed in \code{tissue_map$protein_tissue}.
Any name not in this list will return \code{NA} and issue a warning.}
}
\value{
A character vector of GTEx tissue names.
}
\description{
Converts a vector of tissue names from the Human Protein Atlas (HPA) format
to GTEx-compatible tissue names using the \code{tissue_map} dataset. The tissue
names that each package and database use are different, and this helper function
allows easy conversion for analysis.
If a tissue cannot be converted, it returns "Unknown" and issues a warning.
}
\examples{
/dontrun{
# Example 1: Convert list of simple tissues

example_input <- c("liver", "skin 1", "breast", "appendix")
gtexr_tissues <- convert_to_gtex(tissues = example_input)
gtexr_tissues

# Example 2: Convert list of complex tissues

example_input <- c(
"kidney","lung", "spleen", "liver", "ovary", "testis",
"breast","adrenal gland", "pancreas", "cerebellum")
gtexr_tissues <- convert_to_gtex(tissues = example_input)
gtexr_tissues

}

}
\references{
Lonsdale, J., Thomas, J., Salvatore, M. et al.
The Genotype-Tissue Expression (GTEx) project. Nat Genet 45, 580–585 (2013).
https://doi.org/10.1038/ng.2653

Thul PJ, Lindskog C. The human protein atlas: A spatial map of the human proteome.
Protein Sci. 2018 Jan;27(1):233-244. doi: 10.1002/pro.3307.
Epub 2017 Oct 10. PMID: 28940711; PMCID: PMC5734309.

Tran AN, Dussaq AM, Kennell Jr T, Willey CD, Hjelmeland AB (2019).
“HPAanalyze: an R package that facilitates the retrieval and analysis of the
Human Protein Atlas data.” MC Bioinformatics 20, 463 (2019).
https://doi.org/10.1186/s12859-019-3059-z

Warwick A, Zuckerman B, Ung C, Luben R, Olvera-Barrios A (2025). “gtexr: A
convenient R interface to the Genotype-Tissue Expression (GTEx) Portal API.”
Journal of Open Source So ware, 10(109), 8249. ISSN
2475-9066, doi:10.21105/joss.08249, gigs v0.2.1.
}
