% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Correlation.R
\name{correlation_genes_only}
\alias{correlation_genes_only}
\title{Compute Spearman Correlation Between RNA and Protein Expression for Genes}
\usage{
correlation_genes_only(gene_list)
}
\arguments{
\item{gene_list}{A character vector of gene symbols. Only valid genes will be processed.
Valid gene symbols for the \code{gene_NAMES} argument can be accessed from
\code{gene_symbols_list}.}
}
\value{
A list containing:
\describe{
\item{per_gene_plot}{A \code{ggplot2} object showing Spearman correlation values
for each gene across tissues. Bars are color-coded from red (negative correlation)
to blue (positive correlation).}
}
}
\description{
This function calculates the Spearman correlation between RNA and protein expression levels
for a given list of genes across multiple standard tissues. It also generates a bar plot
visualizing the correlation values for each gene.
}
\details{
The function performs the following steps:
\itemize{
\item Converts gene symbols to GENCODE IDs using \code{get_gtex_gencode_ids}.
\item Queries RNA expression data for standard tissues using \code{gtexr::get_gene_expression}.
\item Aggregates RNA expression by averaging multiple samples per gene per tissue.
\item Downloads and filters normal tissue protein expression data using \code{HPAanalyze::hpaDownload}.
\item Checks for missing gene-tissue combinations in protein data, issuing a warning and filling missing entries with NA.
\item Converts protein expression levels to numeric values using \code{protein_expr_values}.
\item Aligns RNA and protein data and creates a separate tibble for each gene.
\item Computes Spearman correlations between RNA and protein expression across tissues.
\item Generates a bar plot displaying the Spearman correlation for each gene.
\itemize{
\item Plot key - Spearman correlation values are represented by color intensity. High positive correlations are indicated in dark blue,
low correlations around zero are white, and high negative correlations are indicated in dark red.
The shade intensity reflects the magnitude of the correlation.
}
}
}
\note{
ExprCompareR retrieves data from HPA (Thul et al., 2018) and GTEx (Lonsdale et al., 2013)
via packages, HPAanalyze(Tran et al., 2019) and gtexr(Warwick et al., 2025).
Because these functions make external API calls, some operations may take time to complete
typically around 30 seconds to 1 minute.
}
\examples{
\dontrun{

# Example 1: Using valid gene names only

gene_list <- c("MYC", "TP53", "BRCA1", "CRP", "EGFR")
output<- correlation_genes_only(gene_NAMES = gene_list)
output$per_gene_plot

}

}
\references{
Tran AN, Dussaq AM, Kennell Jr T, Willey CD, Hjelmeland AB (2019).
"HPAanalyze: an R package that facilitates the retrieval and analysis of the
Human Protein Atlas data." MC Bioinformatics 20, 463 (2019).
\url{https://doi.org/10.1186/s12859-019-3059-z}

Warwick A, Zuckerman B, Ung C, Luben R, Olvera-Barrios A (2025). "gtexr: A
convenient R interface to the Genotype-Tissue Expression (GTEx) Portal API."
Journal of Open Source Software, 10(109), 8249. ISSN 2475-9066,
\url{doi:10.21105/joss.08249}, gigs v0.2.1.

Wickham H (2016). ggplot2: Elegant Graphics for Data Analysis.
Springer-Verlag New York. ISBN 978-3 319-24277-4, \url{https://ggplot2.tidyverse.org.}

Wickham H, Francois R, Henry L, Muller K, Vaughan D (2025).
dplyr: A Grammar of Data Manipulation. R package version 1.1.4,
\url{https://dplyr.tidyverse.org.}

Wickham H, Henry L (2025). purrr: Functional Programming Tools.
R package version 1.1.0, \url{https://purrr.tidyverse.org.}

Wickham H, Vaughan D, Girlich M (2025). tidyr: Tidy Messy Data.
R package version 1.3.1, \url{https://tidyr.tidyverse.org.}
}
\keyword{internal}
