% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Correlation.R
\name{correlation_genes_tissues}
\alias{correlation_genes_tissues}
\title{Generate Correlation Plots for given list of genes and tissues}
\usage{
correlation_genes_tissues(gene_NAMES, tissue_NAMES, plot_choice)
}
\arguments{
\item{gene_NAMES}{A character vector of gene symbols. Must contain at least five elements if provided.
Valid gene symbols for the \code{gene_NAMES} argument can be accessed from
\code{gene_symbols_list}.}

\item{tissue_NAMES}{A character vector of tissue names. Must contain at least five elements if provided.
Valid tissue names as for the \code{tissue_NAMES} argument can be accessed from
\code{tissue_map$protein_tissue}}

\item{plot_choice}{A character string specifying which plot to generate.
Options are \code{"per_gene"} to visualize correlation for each gene, or \code{"per_tissue"}
to visualize correlation for each tissue.}
}
\value{
Depending on \code{plot_choice}, the function returns:
\describe{
\item{per_gene_plot}{A plot object showing Spearman correlations per gene (if \code{plot_choice = "per_gene"}).}
\item{per_tissue_plot}{A plot object showing Spearman correlations per tissue (if \code{plot_choice = "per_tissue"}).}
}
}
\description{
This function serves as a wrapper to compute and visualize the Spearman correlation
between RNA and protein expression between a user's gene list \strong{AND} tissue list of interest.
Depending on the user's choice, it calls either \code{per_gene_plot} or \code{per_tissue_plot} internally.
\code{per_gene_plot} visualizes the spearman correlation and plot per gene from the input list.
\code{per_tissue_plot} visualizes the spearman correlation and plot per tissue from the input list.
}
\details{
The function validates the inputs before generating plots:
\itemize{
\item If both \code{gene_NAMES} and \code{tissue_NAMES} are NULL, it stops execution and throws an error.
\item If both lists have fewer than five elements, it stops execution and throws an error.
\item If only one list is provided, it stops execution and throws an error.
\item The \code{plot_choice} argument must be either \code{"per_gene"} or \code{"per_tissue"}. Any other value stops execution and throws an error.
}

The function then calls:
\itemize{
\item \code{per_gene_plot(tissue_NAMES, gene_NAMES)} if \code{plot_choice = "per_gene"}.
\item \code{per_tissue_plot(tissue_NAMES, gene_NAMES)} if \code{plot_choice = "per_tissue"}.
}
}
\note{
This function computes correlations for a list of genes \strong{AND} a list of tissues.
If you wish to input  \strong{ONLY} a list of genes \strong{OR} list of tissues, utilize function \code{compute_correlation}, or
see \code{help} documentation \code{?compute_correlation}
}
\examples{
\dontrun{
# Example 1: Generate correlation plot per gene

tissues = c("lung", "spleen", "liver", "ovary", "testis")
genes = c("MYC", "TP53", "BRCA1", "CRP", "EGFR")

result_gene <- correlation_genes_tissues(gene_NAMES = genes, tissue_NAMES = tissues, plot_choice = "per_gene")
result_gene$per_gene_plot

# Example 2: Generate correlation plot per tissue

tissues = c("lung", "spleen", "liver", "ovary", "testis")
genes = c("MYC", "TP53", "BRCA1", "CRP", "EGFR")

result_gene <- correlation_genes_tissues(gene_NAMES = genes,tissue_NAMES = tissues, plot_choice = "per_tissue")
result_gene$per_tissue_plot
}

}
\references{
Tran AN, Dussaq AM, Kennell Jr T, Willey CD, Hjelmeland AB (2019).
"HPAanalyze: an R package that facilitates the retrieval and analysis of the
Human Protein Atlas data." MC Bioinformatics 20, 463 (2019).
\url{https://doi.org/10.1186/s12859-019-3059-z}

Warwick A, Zuckerman B, Ung C, Luben R, Olvera-Barrios A (2025). "gtexr: A
convenient R interface to the Genotype-Tissue Expression (GTEx) Portal API."
Journal of Open Source Software, 10(109), 8249. ISSN 2475-9066,
\url{doi:10.21105/joss.08249}, gigs v0.2.1.

Wickham H (2016). ggplot2: Elegant Graphics for Data Analysis.
Springer-Verlag New York. ISBN 978-3 319-24277-4, \url{https://ggplot2.tidyverse.org.}

Wickham H, Fracois R, Henry L, Muller K, Vaughan D (2025).
dplyr: A Grammar of Data Manipulation. R package version 1.1.4,
\url{https://dplyr.tidyverse.org.}

Wickham H, Henry L (2025). purrr: Functional Programming Tools.
R package version 1.1.0, \url{https://purrr.tidyverse.org.}

Wickham H, Vaughan D, Girlich M (2025). tidyr: Tidy Messy Data.
R package version 1.3.1, \url{https://tidyr.tidyverse.org.}
}
