% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Correlation.R
\name{correlation_tissues_only}
\alias{correlation_tissues_only}
\title{Compute Spearman Correlation Between RNA and Protein Expression for Tissues}
\usage{
correlation_tissues_only(tissue_NAMES)
}
\arguments{
\item{tissue_NAMES}{A character vector of tissue names. Only valid tissues (matching
the standard protein tissue list) will be processed.
Valid tissue names as for the \code{tissue_NAMES} argument can be accessed from
\code{tissue_map$protein_tissue}}
}
\value{
A list containing:
\describe{
\item{per_tissue_plot}{A \code{ggplot2} object showing Spearman correlation values
for each tissue.}
}
}
\description{
This function calculates the Spearman correlation between RNA and protein expression levels
for the top expressed genes in each tissue provided. It generates a bar plot visualizing
the correlation for each tissue.
}
\details{
The function performs the following steps:
\itemize{
\item Downloads and subsets normal tissue protein expression data using \code{HPAanalyze::hpaDownload}.
\item Queries RNA expression data from GTEx for the top expressed genes in each tissue.
\item Filters protein expression data to include only the top expressed genes.
\item Converts protein expression levels to numeric values using \code{protein_expr_values}.
\item Aggregates protein expression by median per gene per tissue.
\item Creates a combined table of RNA and protein expression per tissue.
\item Computes Spearman correlations between RNA and protein expression for each tissue.
\item Generates a bar plot displaying Spearman correlations across tissues.
\itemize{
\item Plot key - Spearman correlation values are represented by color intensity. High positive correlations are indicated in dark blue,
low correlations around zero are white, and high negative correlations are indicated in dark red.
The shade intensity reflects the magnitude of the correlation.
}

}
}
\note{
ExprCompareR retrieves data from HPA (Thul et al., 2018) and GTEx (Lonsdale et al., 2013)
via packages, HPAanalyze(Tran et al., 2019) and gtexr(Warwick et al., 2025).
Because these functions make external API calls, some operations may take time to complete
typically around 30 seconds to 1 minute.
}
\examples{
\dontrun{
# Example 1: Using tissue names only

tissue_list = c("lung", "spleen", "liver", "ovary", "testis")
output<- correlation_tissues_only(tissue_NAMES = tissue_list)
output$per_tissue_plot
}

}
\references{
Tran AN, Dussaq AM, Kennell Jr T, Willey CD, Hjelmeland AB (2019).
"HPAanalyze: an R package that facilitates the retrieval and analysis of the
Human Protein Atlas data." MC Bioinformatics 20, 463 (2019).
\url{https://doi.org/10.1186/s12859-019-3059-z}

Warwick A, Zuckerman B, Ung C, Luben R, Olvera-Barrios A (2025). "gtexr: A
convenient R interface to the Genotype-Tissue Expression (GTEx) Portal API."
Journal of Open Source Software, 10(109), 8249. ISSN 2475-9066,
\url{doi:10.21105/joss.08249}, gigs v0.2.1.

Wickham H (2016). ggplot2: Elegant Graphics for Data Analysis.
Springer-Verlag New York. ISBN 978-3 319-24277-4, \url{https://ggplot2.tidyverse.org.}

Wickham H, Francois R, Henry L, Muller K, Vaughan D (2025).
dplyr: A Grammar of Data Manipulation. R package version 1.1.4,
\url{https://dplyr.tidyverse.org.}

Wickham H, Henry L (2025). purrr: Functional Programming Tools.
R package version 1.1.0, \url{https://purrr.tidyverse.org.}

Wickham H, Vaughan D, Girlich M (2025). tidyr: Tidy Messy Data.
R package version 1.3.1, \url{https://tidyr.tidyverse.org.}
}
\keyword{internal}
