% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Outlier_Detection.R
\name{detect_outliers}
\alias{detect_outliers}
\title{Detects outlier genes present in a given tissue of interest, which have a
large delta between their RNA expression and protein expression values}
\usage{
detect_outliers(input_tissue)
}
\arguments{
\item{input_tissue}{Character String specifying the tissue of interest to analyze.
The name must be present in the given list of tissues:
Valid tissue names are listed in \code{names(tissue_map)}.
Any name not in this list will issue a warning.}
}
\value{
Returns an object of class list with two components:
\itemize{
\item \code{outlier_vector}: A character vector of outlier genes names detected
\item \code{outlier_plot}: A ggplot object highlighting the outlier genes in red
}
}
\description{
The function given a tissue of interest, compares the RNA expression and protein expression
of all genes present in the given tissue, and identifies genes which have a large difference
in their RNA and protein expression. These genes are identified as outlier genes, and are most
likely candidates for post transcriptional modifications. This function returns the list of outlier genes
as well as a plot highlighting the outlier genes.
}
\details{
The function downloads protein expression data from the Human Protein Atlas
using \pkg{HPAanalyze} and RNA expression data from the GTEx portal using \pkg{gtexr}.
It compares both modalities for the specified tissue and flags outlier genes
based on interquartile range (IQR) thresholds for RNA and protein expression.
}
\examples{
\dontrun{
# Example 1: Detect outliers in lung tissue

output <- detect_outliers(input_tissue = "lung")

output$outlier_vector
output$outlier_plot


# Example 2: Detect outlier in the ovary tissue

output <- detect_outliers(input_tissue = "ovary")

output$outlier_vector
output$outlier_plot
}

}
\references{
Warwick A, Zuckerman B, Ung C, Luben R, Olvera-Barrios A (2025). “gtexr: A
convenient R interface to the Genotype-Tissue Expression (GTEx) Portal API.”
Journal of Open Source So ware, 10(109), 8249. ISSN
2475-9066, doi:10.21105/joss.08249, gigs v0.2.1.

Tran AN, Dussaq AM, Kennell Jr T, Willey CD, Hjelmeland AB (2019).
“HPAanalyze: an R package that facilitates the retrieval and analysis of the
Human Protein Atlas data.” BMC

Wickham H (2016). ggplot2: Elegant Graphics for Data Analysis.
Springer-Verlag New York. ISBN 978-3 319-24277-4, https://ggplot2.tidyverse.org.

Wickham H, Henry L (2025). purrr: Functional Programming Tools.
R package version 1.1.0, https://purrr.tidyverse.org/.

Wickham H, Vaughan D, Girlich M (2025). tidyr: Tidy Messy Data.
R package version 1.3.1, https://tidyr.tidyverse.org.

Wickham H, François R, Henry L, Müller K, Vaughan D (2025).
dplyr: A Grammar of Data Manipulation. R package version 1.1.4,
https://dplyr.tidyverse.org.
}
