% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExprRefCompare.R
\name{ExprRefCompare}
\alias{ExprRefCompare}
\alias{ExprRefCompare.seurat}
\alias{ExprRefCompare.Seurat}
\title{ExprRefCompare}
\usage{
ExprRefCompare(object, ...)

\method{ExprRefCompare}{seurat}(seuratObj, accession, GEOentry_index = 1,
  rename_samples = TRUE, do_plot = FALSE, group_by = NULL,
  add_ident = NULL, cor_function_use = cor, platform = NULL,
  probe_set = NULL, annotation_db = NULL, species = "hsapiens", ...)

\method{ExprRefCompare}{Seurat}(object, accession, GEOentry_index = 1,
  rename_samples = TRUE, do_plot = FALSE, group_by = NULL,
  assay = "RNA", add_ident = NULL, cor_function_use = cor,
  platform = NULL, probe_set = NULL, annotation_db = NULL,
  species = "hsapiens", ...)
}
\arguments{
\item{object}{Either a processed Seurat object (with ScaleData run) or
the output from AverageExpression. If a Seurat object is passed,
AverageExpression will be run.}

\item{...}{Additional parameters to pass to the correlation function.}

\item{accession}{A GEO or ArrayExpress accession number (i.e. 'GSE24759')}

\item{GEOentry_index}{Indicate which ExpressionSet to use in a GEO entry
containing multiple ExpressionSets}

\item{rename_samples}{Replace the sample name with that provided in the
ExpressionSet's phenoData@data slot}

\item{do_plot}{Plot the correlation matrix. Default: FALSE}

\item{group_by}{Identifier or meta.data column by which to group the data.
Default: 'ident'}

\item{add_ident}{An additional identifier to use as a grouping variable.
Default: NULL}

\item{cor_function_use}{Function to use to calculate correlation.
Default: stats::cor}

\item{platform}{For ArrayExpress datasets, the Annotation Database
corresponding to the platform used in the assay (i.e. hgu133plus2.db).
If not supplied, will attempt to detect the correct database and retrieve
the names from Biomart (which can be SLOW) Default: NULL}

\item{probe_set}{If it is difficult/impossible to find the correct annotation
package for the assay, you can use biomaRt as a backup by supplying the
probe_set name and species.  For available probesets, run
biomaRt::listAttributes(mart = useMart(mart = "ENSEMBL_MART_ENSEMBL",
dataset = "{species}_gene_ensembl")), where {species} is the same as below.
Default = NULL}

\item{annotation_db}{For entries that do not contain a probeID-to-gene mapping,
it is necessary to supply the name of the annotation package that
corresponds to the assay used in the GEO/AE entry (i.e.
hugene10sttranscriptcluster.db for the [HuGene-1_0-st] Affymetrix Human
Gene 1.0 ST Array [transcript (gene) version].  Default: NULL.}

\item{species}{= Dataset species, in the form of genus initial, species with
no space between.  Default: "hsapiens"}
}
\value{
A matrix of correlation values
}
\description{
Compares the averaged expression of each cluster within a Seurat object
to that of an entry in the Gene Expression Omnibus database or the
ArrayExpress Archive of Functional Genomics Data
}
\details{
Performs a canonical correlation analysis on the two datasets and then
assesses the correlation using the CCA data.
}
