% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_symbols_list}
\alias{gene_symbols_list}
\title{Valid list of gene symbols analysis in ExprCompareR}
\format{
A character vector of length 15313 (one element per gene).
}
\source{
Derived from Human Protein Atlas (\url{https://www.proteinatlas.org/}),
via HPAanalyze
}
\usage{
gene_symbols_list
}
\description{
A character vector of unique valid gene symbols from the HPA normal tissue dataset
used by ExprCompareR.
The symbols can be converted to their respective GENCODE IDs for retrieval
of their RNA expression data.
}
\examples{
data(gene_symbols_list)
length(gene_symbols_list)
head(gene_symbols_list)
\dontrun{
gene_symbols_list
}

}
\references{
Thul PJ, Lindskog C. The human protein atlas: A spatial map of the human proteome.
Protein Sci. 2018 Jan;27(1):233-244. doi: 10.1002/pro.3307.
Epub 2017 Oct 10. PMID: 28940711; PMCID: PMC5734309.

Tran AN, Dussaq AM, Kennell Jr T, Willey CD, Hjelmeland AB (2019).
"HPAanalyze: an R package that facilitates the retrieval and analysis of the
Human Protein Atlas data." MC Bioinformatics 20, 463 (2019).
https://doi.org/10.1186/s12859-019-3059-z
}
\keyword{datasets}
