% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocompare.R
\name{GEOcompare}
\alias{GEOcompare}
\title{GEOcompare}
\usage{
GEOcompare(seuratObj, GEOaccession, do.plot = FALSE, group.by = NULL,
  add.ident = NULL, function.use = cor)
}
\arguments{
\item{seuratObj}{}

\item{GEOaccession}{A GEO accession number (i.e. 'GSE24759')}

\item{do.plot}{Plot the correlation matrix}

\item{group.by}{Identifier or meta.data column by which to group the data. (default: 'ident')}

\item{add.ident}{An additional identifier to use as a grouping variable.}

\item{cor.function.use}{= cor}

\item{method.use}{= 'pearson'}
}
\value{
A matrix of correlation values
}
\description{
Compares the averaged expression of each cluster within a Seurat object
to that of an entry in the Gene Expression Omnibus database.
}
\details{
Performs a canonical correlation analysis on the two datasets and then
assesses the correlation using the CCA data.
}
