% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEOprep.R
\name{GEOprep}
\alias{GEOprep}
\title{GEOprep}
\usage{
GEOprep(GEOaccession, var_list = NULL, index = 1, probe_set = NULL,
  annotation_db = NULL, species = "hsapiens", rename_samples = TRUE)
}
\arguments{
\item{GEOaccession}{A GEO accession number (i.e. 'GSE24759')}

\item{index}{Indicate which ExpressionSet to use in a GEO entry containing
multiple ExpressionSets. Default: 1}

\item{probe_set}{If it is difficult/impossible to find the correct annotation
package for the assay, you can use biomaRt as a backup by supplying the
probe_set name and species.  For available probesets, run
biomaRt::listAttributes(mart = useMart(mart = "ENSEMBL_MART_ENSEMBL",
dataset = "{species}_gene_ensembl")), where {species} is the same as below.
Default = NULL}

\item{annotation_db}{For entries that do not contain a probeID-to-gene mapping,
it is necessary to supply the name of the annotation package that
corresponds to the assay used in the GEO/AE entry (i.e.
hugene10sttranscriptcluster.db for the [HuGene-1_0-st] Affymetrix Human
Gene 1.0 ST Array [transcript (gene) version].  Default: NULL.}

\item{species}{= Dataset species, in the form of genus initial, species with
no space between.  Default: "hsapiens" .}

\item{rename_samples}{Replace the GEO sample name with the title
provided in the ExpressionSet's phenoData@data slot. Default: TRUE}

\item{var.list}{List of variables (i.e. genes) to include.
If provided, data for other variables is discarded. Default: NULL}
}
\value{
A matrix of correlation values
}
\description{
Processes an entry in the Gene Expression Omnibus database for use in
GEOcompare
}
