% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{get_gtex_gencode_ids}
\alias{get_gtex_gencode_ids}
\title{Converts Gene Symbols to GENCODE IDs}
\usage{
get_gtex_gencode_ids(gene_symbols)
}
\arguments{
\item{gene_symbols}{A character vector of gene symbols to query.
Valid symbols are listed in \code{\link{gene_symbols_list}}.
These genes typically have protein expression data but may not always
have corresponding RNA expression data. Missing or unmatched genes will
trigger a warning and be excluded from the output.}
}
\value{
A character vector of corresponding GENCODE IDs.
Genes not found in the GTEx database are omitted with a warning.
}
\description{
The function given a vector of gene symbols, converts them into their
respective GENCODE IDs. This helper function is useful when querying RNA expression
data of a given gene, using the \pkg{gtexr} package, which requires input genes
to be provided as GENCODE IDs.
It serves as an internal utility within the ExprCompareR package, allowing
users to easily map gene symbols (from the Human Protein Atlas) to their
GENCODE equivalents required for GTEx queries.
}
\details{
The function retrieves metadata from the GTEx database via the
\code{gtexr::get_genes()} function and extracts the \code{gencodeId} values
corresponding to the provided gene symbols.
}
\examples{
\dontrun{
# Example 1: Retrieve GENCODE IDs for common genes.

my_genes <- c("MYC", "TP53", "BRCA1")

gencode_ids <- get_gtex_gencode_ids(gene_symbols = my_genes)
gencode_ids


# Example 2: Include a gene symbol which does not exist (will issue a warning).

my_genes <- c("CRP", "EGFR", "UNKNOWN")

gencode_ids <- get_gtex_gencode_ids(gene_symbols = my_genes)
gencode_ids
}

}
\references{
Lonsdale, J., Thomas, J., Salvatore, M. et al.
The Genotype-Tissue Expression (GTEx) project. Nat Genet 45, 580–585 (2013).
https://doi.org/10.1038/ng.2653

Thul PJ, Lindskog C. The human protein atlas: A spatial map of the human proteome.
Protein Sci. 2018 Jan;27(1):233-244. doi: 10.1002/pro.3307.
Epub 2017 Oct 10. PMID: 28940711; PMCID: PMC5734309.

Tran AN, Dussaq AM, Kennell Jr T, Willey CD, Hjelmeland AB (2019).
“HPAanalyze: an R package that facilitates the retrieval and analysis of the
Human Protein Atlas data.” MC Bioinformatics 20, 463 (2019).
https://doi.org/10.1186/s12859-019-3059-z

Warwick A, Zuckerman B, Ung C, Luben R, Olvera-Barrios A (2025). “gtexr: A
convenient R interface to the Genotype-Tissue Expression (GTEx) Portal API.”
Journal of Open Source So ware, 10(109), 8249. ISSN
2475-9066, doi:10.21105/joss.08249, gigs v0.2.1.
}
