% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pathology}
\alias{pathology}
\title{Cancer pathology protein expression data}
\format{
A data frame with 11 variables:
\describe{
\item{ensembl}{Ensembl gene identifier (character).}
\item{gene}{HGNC-approved gene symbol (character).}
\item{cancer}{Name of the cancer type (character).}
\item{high}{Number or proportion of samples with high protein expression.}
\item{medium}{Number or proportion of samples with medium expression.}
\item{low}{Number or proportion of samples with low expression.}
\item{not_detected}{Number or proportion of samples where protein was not detected.}
\item{prognostic_favorable}{Logical or categorical indicator of favorable prognostic association.}
\item{unprognostic_favorable}{Indicator for unprognostic favorable cases.}
\item{prognostic_unfavorable}{Indicator of unfavorable prognostic association.}
\item{unprognostic_unfavorable}{Indicator for unprognostic unfavorable cases.}
}
}
\source{
Data derived from Human Protein Atlas \url{https://www.proteinatlas.org/}
}
\usage{
pathology
}
\description{
A data frame describing protein expression profiles and prognostic
associations across multiple cancer types, as reported in the Human Protein Atlas.
}
\details{
This dataset can be used to explore relationships between protein expression
and cancer prognosis across different tumor types.
The dataset \code{pathology} was derived from the
Human Protein Atlas \code{histology} data, which can be downloaded using:
\preformatted{
downloadedData <- hpaDownload(downloadList = 'histology', version = 'example')
pathology <- downloadedData$pathology
}
}
\examples{
data(pathology)
head(pathology)
names(pathology)
\dontrun{
pathology
}

}
\references{
Thul PJ, Lindskog C. The human protein atlas: A spatial map of the human proteome.
Protein Sci. 2018 Jan;27(1):233-244. doi: 10.1002/pro.3307.
Epub 2017 Oct 10. PMID: 28940711; PMCID: PMC5734309.

Tran AN, Dussaq AM, Kennell Jr T, Willey CD, Hjelmeland AB (2019).
“HPAanalyze: an R package that facilitates the retrieval and analysis of the
Human Protein Atlas data.” MC Bioinformatics 20, 463 (2019).
https://doi.org/10.1186/s12859-019-3059-z
}
\seealso{
\link[HPAanalyze:hpaDownload]{HPAanalyze::hpaDownload}
}
\keyword{datasets}
