% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Correlation.R
\name{per_tissue_plot}
\alias{per_tissue_plot}
\title{Generate Spearman Correlation Plot Per Tissue Across User's Gene List and Tissue List Of Interest}
\usage{
per_tissue_plot(gene_NAMES, tissue_NAMES)
}
\arguments{
\item{gene_NAMES}{A character vector of gene symbols. Must contain at least five elements if provided.
Valid gene symbols for the \code{gene_NAMES} argument can be accessed from
\code{gene_symbols_list}.}

\item{tissue_NAMES}{A character vector of tissue names. Must contain at least five elements if provided.
Valid tissue names as for the \code{tissue_NAMES} argument can be accessed from
\code{tissue_map$protein_tissue}}
}
\value{
A list containing:
\describe{
\item{per_tissue_plot}{A \code{ggplot2} object showing Spearman correlation values
for each tissue across the specified genes. Bars are labeled with correlation values.}
}
}
\description{
This function calculates the Spearman correlation between RNA and protein expression
for the top genes in \code{gene_NAMES} across the specified tissues in \code{tissue_NAMES}.
It produces a bar plot visualizing the correlation for each tissue.
}
\details{
The function performs the following steps:
\itemize{
\item Converts tissue names to GTEx-compatible names using \code{convert_to_gtex}.
\item Converts gene symbols to GENCODE IDs using \code{get_gtex_gencode_ids}.
\item Queries RNA expression data from GTEx for each gene and tissue combination.
\item Aggregates RNA expression by averaging multiple samples per gene per tissue.
\item Downloads normal tissue protein expression data using \code{HPAanalyze::hpaDownload}.
\item Filters protein data to include only the genes and tissues of interest.
\item Handles missing gene-tissue combinations by filling missing entries with \code{NA} and issuing a warning.
\item Converts protein expression levels to numeric using \code{protein_expr_values}.
\item Aggregates protein data by mean per gene per tissue.
\item Computes Spearman correlations between RNA and protein expression for each tissue.
\item Generates a bar plot showing correlations for all tissues.
\itemize{
\item Plot key - Spearman correlation values are represented by color intensity. High positive correlations are indicated in dark blue,
low correlations around zero are white, and high negative correlations are indicated in dark red.
The shade intensity reflects the magnitude of the correlation.
}
}
}
\examples{
\dontrun{
# Example 2: Generate correlation plot per tissue


result_gene <- per_tissue_plot(gene_NAMES = genes,tissue_NAMES = tissues)
result_gene$per_tissue_plot
}

}
\references{
Müller K, Wickham H (2025). tibble: Simple Data Frames.
R package version 3.3.0, \url{https://tibble.tidyverse.org/.}

Tran AN, Dussaq AM, Kennell Jr T, Willey CD, Hjelmeland AB (2019).
"HPAanalyze: an R package that facilitates the retrieval and analysis of the
Human Protein Atlas data." MC Bioinformatics 20, 463 (2019).
\url{https://doi.org/10.1186/s12859-019-3059-z}

Warwick A, Zuckerman B, Ung C, Luben R, Olvera-Barrios A (2025). "gtexr: A
convenient R interface to the Genotype-Tissue Expression (GTEx) Portal API.
Journal of Open Source Software, 10(109), 8249. ISSN 2475-9066,
\url{doi:10.21105/joss.08249}, gigs v0.2.1.

Wickham H (2016). ggplot2: Elegant Graphics for Data Analysis.
Springer-Verlag New York. ISBN 978-3 319-24277-4,
\url{https://ggplot2.tidyverse.org.}

Wickham H, Francois R, Henry L, Muller K, Vaughan D (2025).
dplyr: A Grammar of Data Manipulation. R package version 1.1.4,
\url{https://dplyr.tidyverse.org.}

Wickham H, Henry L (2025). purrr: Functional Programming Tools.
R package version 1.1.0, \url{https://purrr.tidyverse.org/.}

Wickham H, Vaughan D, Girlich M (2025). tidyr: Tidy Messy Data.
R package version 1.3.1, \url{https://tidyr.tidyverse.org.}
}
\keyword{internal}
