% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeuratCompareR.R
\name{SeuratCompareR}
\alias{SeuratCompareR}
\title{SeuratCompareR}
\usage{
SeuratCompareR(first.obj, ...)
}
\arguments{
\item{first.obj}{Either a processed Seurat object (with ScaleData run) or
the output from AverageExpression.  If a Seurat object is passed,
AverageExpression will be run.}

\item{...}{Additional parameters to pass to the correlation function.}

\item{second.obj}{A second (or the same) processed Seurat object or the
output from AverageExpression for such an object.}

\item{group.by}{Identifier or meta.data column by which to group the data.
Default: 'ident'}

\item{add.ident}{An additional identifier to use as a grouping variable.
Default: NULL}

\item{cor.function.use}{Function to use to calculate correlation.
Default: stats::cor}

\item{do.plot}{Plot the correlation matrix. Default: FALSE)}
}
\value{
A matrix of correlation values
}
\description{
Uses Canonical Correlation Analysis to compare the expression profile of the
identity classes of a Seurat object to those of a second Seurat object
}
