% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tissue_list_protein}
\alias{tissue_list_protein}
\title{Default list of protein tissues used for correlation analysis}
\format{
A character vector of length 10 (one element per tissue)
}
\source{
Internal dataset created for ExprCompareR package.
Data derived from Human Protein Atlas (\url{https://www.proteinatlas.org/}),
via HPAanalyze
}
\usage{
tissue_list_protein
}
\description{
A character vector of the default tissues that function correlation_per_gene
uses to analyze RNA vs protein expression when the input is a list of genes.
These tissue names follow the format of HPAanalyze and can only be used to query
}
\examples{
data(tissue_list_protein)
length(tissue_list_protein)
head(tissue_list_protein)
\dontrun{
tissue_list_protein
}

}
\references{
Thul PJ, Lindskog C. The human protein atlas: A spatial map of the human proteome.
Protein Sci. 2018 Jan;27(1):233-244. doi: 10.1002/pro.3307.
Epub 2017 Oct 10. PMID: 28940711; PMCID: PMC5734309.

Tran AN, Dussaq AM, Kennell Jr T, Willey CD, Hjelmeland AB (2019).
"HPAanalyze: an R package that facilitates the retrieval and analysis of the
Human Protein Atlas data." MC Bioinformatics 20, 463 (2019).
https://doi.org/10.1186/s12859-019-3059-z
}
\keyword{datasets}
